+++ src/routes/(app)/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import Chat from '$lib/components/chat/Chat.svelte';
 	import Help from '$lib/components/layout/Help.svelte';
-	import Banner from '$lib/components/sponsor_banner/Banner.svelte';
+	import Banner from '$lib/components/legal_disclaimer/Banner.svelte';
 </script>
 
 <Help />
